viewport: shots take the SHOT's aspect; a piloted constrained camera keeps its authored one (plan item 15, audit C2b.2)
WHAT THE RIG SAID (1600x1000 vs 1920x1080 Xvfb, the scripting.e2e.particles
scene): with the pose fixed by editor.setCamera the two displays' shots are
identical outside the particle plume, and the plume differs just as much between
two processes on ONE display (emission is random per process) - the shot path
never read the window. What differs is the POSE F produces: focusSelection
backs off until the subject fills the viewport's rendered angle, and above
16:9 that angle depends on the window (z 2.867 vs 3.369, 17.5%). That is the
framing hold working as designed; a framed-for-the-window pose cannot be
window-independent without F overflowing wide viewports again (717f1453). So
the "same editor.screenshot on two displays" contract holds for a fixed pose,
and ui.shot_aspect pins exactly that; the F case is recorded as by-design in
takeScreenshot's comment and in the lane report.

THE AUDIT'S FIX, AND WHY IT WAS NOT ENOUGH. takeScreenshot swapped the camera
to the shot's aspect only when a gizmo was drawn; it now does so for the whole
shot, always - which changes no scene pixel (the engine resolves a free
camera's frame from the shot's own target) but keeps every document-side reader
(gizmo size, projMatrix) describing the picture being taken. Except, and this
was a defect: a piloted camera that CONSTRAINS its aspect renders letterboxed
at its AUTHORED aspect in any view, the shot included; the old swap wrote the
shot's shape over it, and a 640x480 shot of a 2.39 camera with a selection had
no bars at all.

THE DEFECT FOUND ON THE WAY (the bigger one): nothing kept a piloted
constrained camera's authored aspect. Gizmo::setPickView (per frame since
smoke S15) and ScenePicker::screenSegment (every hover and pick) set the
camera's aspect from the WIDGET, so on the first frame of piloting a 2.39
camera became the viewport's 1.87: the on-screen letterbox vanished, a save
would have written the viewport's shape into the camera, and shots
letterboxed at the wrong aspect. Every pick path now goes through
pictureRect() - the widget, or the letterbox's inner rectangle
(chain::letterboxRect's arithmetic in pixels) while piloting a constrained
camera - so rays and the gizmo's pixel frame unproject at the aspect that is on
screen and the aspect they set is the one already there. screenSegment takes a
FRACTIONAL size for the same reason (a letterboxed picture is not a whole
number of pixels tall; rounding nudged 2.39 to 2.3887 on every hover).

Test: ui.shot_aspect (MCP harness, app.resizeWindow): the same pose and
selection photographed at a 2.63-aspect and a 1.52-aspect viewport (both sides
of the 16:9 hold) - 0 pixels differ; piloting + a pick + a gizmo hit test
leave the authored 2.39 exact (1e-5); the shot is letterboxed to 267 rows
(640/2.39). Before this commit (same test, viewport files reverted): aspect
1.5146 (the viewport's), letterbox rows 0..479 (none) - 2 failures; the
pixel-identity half passes before too - it pins the contract, it does not
convict a change. On-screen capture after: bars present while piloting
(spikes/l11-a5b-remainder/pilot.png).

+++ src/viewport/scenepicker.h
@@ -28,8 +28,13 @@ class ScenePicker
 public:
     /// World-space segment through a viewport pixel, from the near to the far plane,
     /// for a document camera drawn at `viewportWidth` x `viewportHeight`.
-    static void screenSegment(iris::CameraNodePtr camera, int viewportWidth, int viewportHeight,
-                              const QPointF &point, iris::Vec3 &segStart, iris::Vec3 &segEnd);
+    /// The size is FRACTIONAL on purpose: it sets the camera's aspect, and a
+    /// letterboxed picture (EngineSceneViewport::pictureRect) is not a whole
+    /// number of pixels tall — rounding it would nudge an authored 2.39 to
+    /// 2.3887 on every hover.
+    static void screenSegment(iris::CameraNodePtr camera, qreal viewportWidth,
+                              qreal viewportHeight, const QPointF &point,
+                              iris::Vec3 &segStart, iris::Vec3 &segEnd);
 
     /// Every hit along the segment, unsorted. Meshes are tested against their
     /// triangles in local space; lights, decals and CAMERAS as
